[Add] status bar for pick up
